requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.2
python-decouple==3.0
static3==0.6.1
gunicorn==19.4.1
psycopg2==2.6.1
requests==2.9.1
djangorestframework==3.3.1







